- further improvements for stereo_pose_refinement - THIS CODE MIGHT NOT COMPILE! -> Use the 2nd before the last version git-svn-id: https://redmine.ait.ac.at/sas-svn/navigation/trunk/matching@418 5a9f4643-6f94-47ed-9025-08a63a7baffb

+++ source/poselib/include/poselib/pose_helper.h
@@ -163,5 +163,5 @@ double POSELIB_API getSampsonL2Error(cv::InputArray E, cv::InputArray x1, cv::In
 //Calculates the Sampson L2 error for 1 correspondence
 double POSELIB_API getSampsonL2Error(Eigen::Matrix3d E, Eigen::Vector3d x1, Eigen::Vector3d x2);
 //Checks for a given vector of error values if they are inliers or not in respect to threshold th.
-unsigned int POSELIB_API getInlierMask(std::vector<double> error, double th, cv::Mat & mask);
+size_t POSELIB_API getInlierMask(std::vector<double> error, double th, cv::Mat & mask);
 }
